ext/x25519_ref10: Consolidate all field element code into fe.c
This should improve build times
