--- ext/x25519_ref10/fe_invert.c
@@ -1,14 +1,0 @@
-#include "fe.h"
-
-void fe_invert(fe out,fe z)
-{
-  fe t0;
-  fe t1;
-  fe t2;
-  fe t3;
-  int i;
-
-#include "pow225521.h"
-
-  return;
-}
